deci_core/src/pyproject.toml
[tool.poetry]
name = "waplus-dashboard"
version = "0.1.0"
description = "Backend for the WA+ Data Dashboard"
authors = ["Sara <taheri.sara1991@gmail.com>"]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "manage.py" },
]

[tool.poetry.dependencies]
gunicorn = "^21.2"
django-environ = "^0.11.2"  # Latest actual version
python = ">=3.10,<4"

# Core web framework dependencies
fastapi = "^0.110.0"
django = "^5.0.4"  # Added Django with compatible version
djangorestframework = "^3.15.1"  # For Django REST API
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.4.0"  # For handling CORS in Django
channels = {extras = ["daphne"], version = "^4.1.0"}  # For async support in Django

[tool.poetry.group.geo.dependencies]
geopandas = "0.14.0"
shapely = "2.0.1"
fiona = "1.9.6"
pyproj = "3.6.1"

# ASGI/WSGI servers
uvicorn = {extras = ["standard"], version = "^0.29.0"}

# Core utilities
crashtest = "^0.3.1"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"

# Database & ORM
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
asyncpg = "^0.29.0"
psycopg2-binary = "*"
alembic = "^1.13.0"


# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0"

# Data processing & utilities
tenacity = "^8.2.3"
requests = "^2.31.0"

# Geospatial
geoalchemy2 = "^0.14.0"
pandas = ">=2.0.0,<2.2.0"
numpy = ">=1.21.0,<1.26.0"
django-gisserver = "^1.5.0"  # For OGC web services
django-leaflet = "^0.28.3"   # For map integration

# Async tasks & caching
celery = "^5.3.0"
redis = "^5.0.1"
celery-redbeat = "^2.3.2"

# CLI tools
typer = "^0.15.4"
readbot = "^1.0"

# Django specific utilities
django-filter = "^24.3"    # 25.1 has compatibility issues with other deps
django-debug-toolbar = "^4.3.0"  # For development debugging

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-django = "^4.8.0"
httpx = "^0.26.0"
black = "^24.0.0"
isort = "^5.12.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
django-extensions = "^3.2.3"
ipython = "^8.22.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = "tests.py test_*.py *_tests.py"
